refactor: rename the mod
