+++ src/utils/mod.rs
@@ -2,7 +2,7 @@ use bnf::{Grammar, Term};
 use std::collections::HashSet;
 
 pub mod builder;
-pub mod first;
+pub mod firstv1;
 pub mod follow;
 
 pub fn symbols(grammar: &Grammar) -> HashSet<&Term> {
